storage: move all helpers to pytest-simcore
