+++ services/storage/tests/unit/test_rest.py
@@ -13,6 +13,8 @@ import pytest
 import simcore_service_storage.meta
 from aiohttp import web
 from aiohttp.test_utils import TestClient
+from pytest_simcore.helpers.utils_assert import assert_status
+from pytest_simcore.helpers.utils_project import clone_project_data
 from simcore_service_storage.access_layer import AccessRights
 from simcore_service_storage.app_handlers import HealthCheck
 from simcore_service_storage.constants import APP_CONFIG_KEY, SIMCORE_S3_ID
